Move core_vec(i) to gui
