+++ python/taichi/tools/util.py
@@ -6,60 +6,6 @@ from colorama import Fore, Style
 from taichi._lib import core as _ti_core
 
 
-def core_veci(*args):
-    if isinstance(args[0], _ti_core.Vector2i):
-        return args[0]
-    if isinstance(args[0], _ti_core.Vector3i):
-        return args[0]
-    if isinstance(args[0], tuple):
-        args = tuple(*args)
-    if len(args) == 2:
-        return _ti_core.Vector2i(int(args[0]), int(args[1]))
-    if len(args) == 3:
-        return _ti_core.Vector3i(int(args[0]), int(args[1]), int(args[2]))
-    if len(args) == 4:
-        return _ti_core.Vector4i(int(args[0]), int(args[1]), int(args[2]),
-                                 int(args[3]))
-    assert False, type(args[0])
-
-
-def core_vec(*args):
-    if isinstance(args[0], _ti_core.Vector2f):
-        return args[0]
-    if isinstance(args[0], _ti_core.Vector3f):
-        return args[0]
-    if isinstance(args[0], _ti_core.Vector4f):
-        return args[0]
-    if isinstance(args[0], _ti_core.Vector2d):
-        return args[0]
-    if isinstance(args[0], _ti_core.Vector3d):
-        return args[0]
-    if isinstance(args[0], _ti_core.Vector4d):
-        return args[0]
-    if isinstance(args[0], tuple):
-        args = tuple(*args)
-    if _ti_core.get_default_float_size() == 4:
-        if len(args) == 2:
-            return _ti_core.Vector2f(float(args[0]), float(args[1]))
-        if len(args) == 3:
-            return _ti_core.Vector3f(float(args[0]), float(args[1]),
-                                     float(args[2]))
-        if len(args) == 4:
-            return _ti_core.Vector4f(float(args[0]), float(args[1]),
-                                     float(args[2]), float(args[3]))
-        assert False, type(args[0])
-    else:
-        if len(args) == 2:
-            return _ti_core.Vector2d(float(args[0]), float(args[1]))
-        if len(args) == 3:
-            return _ti_core.Vector3d(float(args[0]), float(args[1]),
-                                     float(args[2]))
-        if len(args) == 4:
-            return _ti_core.Vector4d(float(args[0]), float(args[1]),
-                                     float(args[2]), float(args[3]))
-        assert False, type(args[0])
-
-
 # The builtin `warnings` module is unreliable since it may be suppressed
 # by other packages such as IPython.
 def warning(msg, warning_type=UserWarning, stacklevel=1):
