+++ tests/python/test_mpm88.py
@@ -78,9 +78,6 @@ def run_mpm88_test():
             J[p] *= 1 + dt * new_C.trace()
             C[p] = new_C
 
-    # gui = ti._lib.core.GUI("MPM88", ti.core_veci(512, 512))
-    # canvas = gui.get_canvas()
-
     for i in range(n_particles):
         x[i] = [i % N / N * 0.4 + 0.2, i / N / N * 0.4 + 0.05]
         v[i] = [0, -3]
